requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pinecone-client
llama-cpp-python
sentence-transformers
python-dotenv
numpy
requests
